Finished adding the ignore ssl issues functionality - needs docs.
