+++ http-builder-ng-core/src/main/java/groovyx/net/http/HttpObjectConfig.java
@@ -76,10 +76,18 @@ public interface HttpObjectConfig extends HttpConfig {
          */
         void setSslContext(SSLContext val);
 
-        // FIXME: document
+        /**
+         * Specifies the {@link HostnameVerifier} to be used by the configured client (related to SSL).
+         *
+         * @param verifier the hostname verifier
+         */
         void setHostnameVerifier(HostnameVerifier verifier);
 
-        // FIXME: document
+        /**
+         * Used to retrieve the {@link HostnameVerifier} configured for the client.
+         *
+         * @return the configured {@link HostnameVerifier}
+         */
         HostnameVerifier getHostnameVerifier();
 
         /**
@@ -96,7 +104,7 @@ public interface HttpObjectConfig extends HttpConfig {
          * [source,groovy]
          * ----
          * long elapsed = configure {
-         *      request.uri = 'https://localhost:10101/foo'
+         *      request.uri = 'http://localhost:10101/foo'
          *      execution.interceptor(GET) { ChainedHttpConfig cfg, Function<ChainedHttpConfig, Object> fx ->
          *          long started = System.currentTimeMillis()
          *          fx.apply(cfg)
