+++ http-builder-ng-core/src/test/groovy/groovyx/net/http/EncodersSpec.groovy
@@ -79,7 +79,7 @@ class EncodersSpec extends Specification {
 
         expect:
         http.post {
-            request.uri = "${ersatzServer.serverUrl}/multi"
+            request.uri = "${ersatzServer.httpUrl}/multi"
             request.body = MultipartContent.multipart {
                 part 'filea', fileA.name, TEXT[0], fileA.toPath()
                 part 'fileb', fileB.name, TEXT[0], fileB.toPath()
@@ -101,7 +101,7 @@ class EncodersSpec extends Specification {
 
         expect:
         http.post {
-            request.uri = "${ersatzServer.serverUrl}/multi"
+            request.uri = "${ersatzServer.httpUrl}/multi"
             request.body = MultipartContent.multipart {
                 part 'filea', fileA.name, TEXT[0], fileA.bytes
                 part 'fileb', fileB.name, TEXT[0], fileB.bytes
@@ -123,7 +123,7 @@ class EncodersSpec extends Specification {
 
         expect:
         http.post {
-            request.uri = "${ersatzServer.serverUrl}/multi"
+            request.uri = "${ersatzServer.httpUrl}/multi"
             request.body = MultipartContent.multipart {
                 part 'filea', fileA.name, TEXT[0], fileA.text
                 part 'fileb', fileB.name, TEXT[0], fileB.text
@@ -145,7 +145,7 @@ class EncodersSpec extends Specification {
 
         expect:
         http.post {
-            request.uri = "${ersatzServer.serverUrl}/multi"
+            request.uri = "${ersatzServer.httpUrl}/multi"
             request.body = MultipartContent.multipart {
                 part 'filea', fileA.name, TEXT[0], fileA.newInputStream()
                 part 'fileb', fileB.name, TEXT[0], fileB.newInputStream()
